Fixed Error related to logging (backport from matrix)

+++ resources/libs/common/logging.py
@@ -83,7 +83,7 @@ def check_log():
     elif CONFIG.CLEANWIZLOGBY == '1':  # By Size
         maxsize = CONFIG.MAXWIZSIZE[int(float(CONFIG.CLEANSIZE))]*1024
         if os.path.getsize(CONFIG.WIZLOG) >= maxsize:
-            start = len(lines)/2
+            start = int(len(lines)/2)
             newfile = lines[start:]
             tools.write_to_file(CONFIG.WIZLOG, '\n'.join(newfile))
     elif CONFIG.CLEANWIZLOGBY == '2':  # By Lines
